+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to Mapbox Directions for Swift
 
+## 2.8.0
+
+### Packaging
+
+* This library now requires a minimum deployment target of iOS 12.0 or above, macOS 10.14.0 or above, tvOS 12.0 or above, or watchOS 5.0 or above. Older operating system versions are no longer supported. ([#736](https://github.com/mapbox/mapbox-directions-swift/pull/736))
+
 ## 2.7.0
 
 ### Packaging